Update how the website is updated.
There is a new and improved website_updater.py.

I have made a new people section with a nicer formatting. It includes little icons to link people's email etc.
The section is now constructed from the _basefiles/people.json file.

I have also fixed the pub section for papers that are accepted but not published yet.